requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-chroma
chromadb
PyMuPDF
pydantic 